port to BenchmarkDotNet
